Add min_hour and max_hour arguments.
These are used as the minimum and maximum acceptable hour selections
(as you would expect).
They are only usable for 24 hour, and will be ignored with the 12
hour clock setting, as that would require javascript to change the dropdowns
on am and pm changes.

Also a few changes to make the spacing a bit more pythonic.